Delete old CRM and unload chunk when no longer visible
