+++ src/binstarjs03.AerialOBJ.WpfApp/binstarjs03.AerialOBJ.WpfApp.csproj
@@ -18,14 +18,6 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-	  <Compile Remove="ChunkRegionManager.cs" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <None Include="ChunkRegionManager.cs" />
-	</ItemGroup>
-
-	<ItemGroup>
 	  <PackageReference Include="Microsoft.Xaml.Behaviors.Wpf" Version="1.1.39" />
 	</ItemGroup>
